Updates based on feedback
1) Created a new test to assert whether the nodes visited by dfs and bfs are same.

2) Its not possible to run all the fuzz targets at once, so added another comment in the contributing file that will list all the fuzz targets.
